Closes #56: refactoring of argument parsers out of download.cli and into their specific modules
